feat(cms-sections): support multivariate string/message variants
Extract shared variant wrapper logic into MultivariateFieldWrapper and
add MultivariateStringField for message.ts multivariate types. Previously
Message fields fell through to the raw AnyOfField renderer.
